Reuse ConnectionManager in compiler module
