+++ gradle/libs.versions.toml
@@ -41,7 +41,7 @@ jgrapht = { module = "org.jgrapht:jgrapht-core", version = "1.5.1" }
 truth = { module = "com.google.truth:truth", version = "1.1.3" }
 turbine = { module = "app.cash.turbine:turbine", version = "0.7.0" }
 bugsnag = { module = "com.bugsnag:bugsnag", version = "3.6.3" }
-picnic = { module = "com.jakewharton.picnic:picnic", version = "0.5.0" }
+picnic = { module = "com.jakewharton.picnic:picnic", version = "0.6.0" }
 
 intellij-java = { module = "com.jetbrains.intellij.java:java-psi", version.ref = "idea" }
 intellij-core = { module = "com.jetbrains.intellij.platform:core-impl", version.ref = "idea" }
