+++ sqldelight-idea-plugin/src/main/resources/META-INF/plugin.xml
@@ -108,8 +108,6 @@
         serviceImplementation="app.cash.sqldelight.intellij.ProjectService"/>
     <projectService serviceInterface="app.cash.sqldelight.intellij.run.SqlDelightStatementExecutor"
         serviceImplementation="app.cash.sqldelight.intellij.run.SqlDelightStatementExecutorImpl"/>
-    <projectService serviceInterface="app.cash.sqldelight.intellij.run.ConnectionManager"
-        serviceImplementation="app.cash.sqldelight.intellij.run.ConnectionManagerImpl"/>
 
     <highlightVisitor implementation="app.cash.sqldelight.intellij.lang.SqlDelightHighlightVisitor"/>
